Completed XML documentation. Moved the nested language class into a new file as LanguageHelper for public use.
